Hacked together support for localization on Windows

+++ generate-locales.sh
@@ -1,12 +1,22 @@
 #!/bin/bash
 
+# Constants
 GETTEXT_DOMAIN="syncthing-gtk"
 MESSAGES_PO="messages.po"
+
+if [[ $(uname) == *"_NT"* ]] ; then
+	# MinGW can't do most of this
+	for lang in locale/* ; do
+		filename=${lang}/LC_MESSAGES/${GETTEXT_DOMAIN}.po
+		mo=${lang}/LC_MESSAGES/${GETTEXT_DOMAIN}.mo
+		msgfmt ${filename} -o ${mo}
+	done
+	exit 0
+fi
 
 # Generate messages.po
 [ -e ${MESSAGES_PO} ] && rm ${MESSAGES_PO}
 xgettext -e syncthing_gtk/*.py *.glade
-
 
 for lang in locale/* ; do
 	filename=${lang}/LC_MESSAGES/${GETTEXT_DOMAIN}.po
